Login should work now..

+++ client/core/src/com/boxtrotstudio/ghost/client/network/packets/SessionPacket.kt
@@ -10,7 +10,7 @@ class SessionPacket : Packet<PlayerClient>() {
         val session : String = args[0] as String;
 
         write(0x00.toByte())
-        write(session.length.toByte())
+        write(session.length.toShort())
         write(session)
 
         if (args.size == 2) {
